Feat/ scale, position등 세부 조정 전
